+++ app/vmalert/notifier/alertmanager_test.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/promauth"
+	"github.com/VictoriaMetrics/VictoriaMetrics/lib/promrelabel"
 )
 
 func TestAlertManager_Addr(t *testing.T) {
@@ -76,12 +77,33 @@ func TestAlertManager_Send(t *testing.T) {
 			if a[0].EndAt.IsZero() {
 				t.Fatalf("expected non-zero end time")
 			}
+			if len(a[0].Labels) != 1 {
+				t.Fatalf("expected 1 labels got %d", len(a[0].Labels))
+			}
+			if len(a[0].Annotations) != 2 {
+				t.Fatalf("expected 2 annotations got %d", len(a[0].Annotations))
+			}
 			if r.Header.Get(headerKey) != "bar" {
 				t.Fatalf("expected header %q to be set to %q; got %q instead", headerKey, headerValue, r.Header.Get(headerKey))
 			}
 		case 3:
-			if r.Header.Get(headerKey) != headerValue {
-				t.Fatalf("expected header %q to be set to %q; got %q instead", headerKey, headerValue, r.Header.Get(headerKey))
+			var a []struct {
+				Labels map[string]string `json:"labels"`
+			}
+			if err := json.NewDecoder(r.Body).Decode(&a); err != nil {
+				t.Fatalf("can not unmarshal data into alert %s", err)
+			}
+			if len(a) != 1 {
+				t.Fatalf("expected 1 alert in array got %d", len(a))
+			}
+			if len(a[0].Labels) != 3 {
+				t.Fatalf("expected 3 labels got %d", len(a[0].Labels))
+			}
+			if a[0].Labels["env"] != "prod" {
+				t.Fatalf("expected env label to be prod during relabeling, got %s", a[0].Labels["env"])
+			}
+			if r.Header.Get(headerKey) != "bar" {
+				t.Fatalf("expected header %q to be set to %q; got %q instead", headerKey, "bar", r.Header.Get(headerKey))
 			}
 		}
 	})
@@ -95,31 +117,58 @@ func TestAlertManager_Send(t *testing.T) {
 		},
 		Headers: []string{fmt.Sprintf("%s:%s", headerKey, headerValue)},
 	}
+	parsedConfigs, err := promrelabel.ParseRelabelConfigsData([]byte(`
+- action: drop
+  if: '{tenant="0"}'
+  regex: ".*"
+- target_label: "env"
+  replacement: "prod"
+  if: '{tenant="1"}'
+`))
+	if err != nil {
+		t.Fatalf("unexpected error when parse relabeling config: %s", err)
+	}
 	am, err := NewAlertManager(srv.URL+alertManagerPath, func(alert Alert) string {
 		return strconv.FormatUint(alert.GroupID, 10) + "/" + strconv.FormatUint(alert.ID, 10)
-	}, aCfg, nil, 0)
+	}, aCfg, parsedConfigs, 0)
 	if err != nil {
 		t.Fatalf("unexpected error: %s", err)
 	}
-	if err := am.Send(context.Background(), []Alert{{}, {}}, nil); err == nil {
+
+	if err := am.Send(context.Background(), []Alert{{Labels: map[string]string{"a": "b"}}}, nil); err == nil {
 		t.Fatalf("expected connection error got nil")
 	}
-	if err := am.Send(context.Background(), []Alert{}, nil); err == nil {
+
+	if err := am.Send(context.Background(), []Alert{{Labels: map[string]string{"a": "b"}}}, nil); err == nil {
 		t.Fatalf("expected wrong http code error got nil")
 	}
+
 	if err := am.Send(context.Background(), []Alert{{
 		GroupID:     0,
 		Name:        "alert0",
 		Start:       time.Now().UTC(),
 		End:         time.Now().UTC(),
-		Annotations: map[string]string{"a": "b", "c": "d", "e": "f"},
+		Labels:      map[string]string{"alertname": "alert0"},
+		Annotations: map[string]string{"a": "b", "c": "d"},
 	}}, map[string]string{headerKey: "bar"}); err != nil {
 		t.Fatalf("unexpected error %s", err)
 	}
-	if c != 2 {
-		t.Fatalf("expected 2 calls(count from zero) to server got %d", c)
-	}
-	if err := am.Send(context.Background(), nil, map[string]string{headerKey: headerValue}); err != nil {
+
+	if err := am.Send(context.Background(), []Alert{
+		// drop tenant0 alert message during relabeling
+		{
+			Name:   "alert1",
+			Labels: map[string]string{"rule": "test", "tenant": "0"},
+		},
+		{
+			Name:   "alert2",
+			Labels: map[string]string{"rule": "test", "tenant": "1"},
+		},
+	}, map[string]string{headerKey: "bar"}); err != nil {
 		t.Fatalf("unexpected error %s", err)
+	}
+
+	if c != 3 {
+		t.Fatalf("expected 3 calls(count from zero) to server got %d", c)
 	}
 }